+++ people-and-teams/profilecard/src/types.ts
@@ -425,6 +425,7 @@ export type ProfileCardErrorType = {
 
 export interface ProfileClientOptions {
   url: string;
+  teamsUseV2?: boolean;
   gatewayGraphqlUrl?: string;
   cacheSize?: number;
   cacheMaxAge?: number;