+++ platform/feature-flags/src/registry.ts
@@ -10,12 +10,24 @@ const pkgName = '@atlaskit/platform-feature-flags';
 
 const registry: FeatureFlagRegistry = {};
 
+// FF global overrides can be configured by test runners or Storybook
+const ENV_ENABLE_PLATFORM_FF =
+  // Use global "process" variable and process.env['FLAG_NAME'] syntax, so it can be replaced by webpack DefinePlugin
+  typeof process !== 'undefined'
+    ? process.env['ENABLE_PLATFORM_FF'] === 'true'
+    : false;
+
 // STORYBOOK_ENABLE_PLATFORM_FF is included as storybook only allows env vars prefixed with STORYBOOK
 // https://github.com/storybookjs/storybook/issues/12270
-const IN_TESTS_ENABLE_PLATFORM_FF =
-  (globalThis?.process?.env?.ENABLE_PLATFORM_FF === 'true' ||
-    globalThis?.process?.env?.STORYBOOK_ENABLE_PLATFORM_FF === 'true') ??
-  false;
+
+const ENV_STORYBOOK_ENABLE_PLATFORM_FF =
+  // Use global "process" variable and process.env['FLAG_NAME'] syntax, so it can be replaced by webpack DefinePlugin
+  typeof process !== 'undefined'
+    ? process.env['STORYBOOK_ENABLE_PLATFORM_FF'] === 'true'
+    : false;
+
+const ENABLE_GLOBAL_PLATFORM_FF_OVERRIDE =
+  ENV_ENABLE_PLATFORM_FF || ENV_STORYBOOK_ENABLE_PLATFORM_FF;
 
 // A list of property keys that are not allowed to exist on the store proxy object.
 const IN_TESTS_UNDEFINED_KEYS = [
@@ -58,7 +70,7 @@ const store = new Proxy(registry, {
   },
 
   get(target, key: string): boolean | undefined {
-    if (IN_TESTS_ENABLE_PLATFORM_FF) {
+    if (ENABLE_GLOBAL_PLATFORM_FF_OVERRIDE) {
       debug(
         '[%s]: The feature flags were enabled while running tests. The flag "%s" will be always enabled.',
         pkgName,